+++ R/RStudio.R
@@ -1,6 +1,6 @@
 # @file RStudio.R
 #
-# Copyright 2019 Observational Health Data Sciences and Informatics
+# Copyright 2020 Observational Health Data Sciences and Informatics
 #
 # This file is part of DatabaseConnector
 # 
@@ -196,10 +196,14 @@ getServer <- function(connection) {
 }
 
 getServer.default <- function(connection) {
-  databaseMetaData <- rJava::.jcall(connection@jConnection,
-                                    "Ljava/sql/DatabaseMetaData;",
-                                    "getMetaData")
-  url <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getURL")
+  if (connection@dbms == "hive") {
+    url <- connection@url
+  } else {
+    databaseMetaData <- rJava::.jcall(connection@jConnection,
+                                      "Ljava/sql/DatabaseMetaData;",
+                                      "getMetaData")
+    url <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getURL")
+  }
   server <- urltools::url_parse(url)$domain
   return(server)
 }
@@ -216,8 +220,13 @@ compileReconnectCode.default <- function(connection) {
   databaseMetaData <- rJava::.jcall(connection@jConnection,
                                     "Ljava/sql/DatabaseMetaData;",
                                     "getMetaData")
-  url <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getURL")
-  user <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getUserName")
+  if (connection@dbms == 'hive') {
+    url <- connection@url
+    user <- connection@user
+  } else {
+    url <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getURL")
+    user <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getUserName")
+  }
   code <- sprintf("library(DatabaseConnector)\ncon <- connect(dbms = \"%s\", connectionString = \"%s\", user = \"%s\", password = password)",
                   connection@dbms,
                   url,
@@ -248,7 +257,7 @@ getSchemaNames.default <- function(conn, catalog = NULL) {
   schemas <- character()
   while (rJava::.jcall(resultSet, "Z", "next")) {
     thisCatalog <- rJava::.jcall(resultSet, "S", "getString", "TABLE_CATALOG")
-    if (is.null(thisCatalog) || (!is.null(catalog) && thisCatalog == catalog)) {
+    if (is.jnull(thisCatalog) || (!is.jnull(catalog) && thisCatalog == catalog)) {
       schemas <- c(schemas, rJava::.jcall(resultSet, "S", "getString", "TABLE_SCHEM"))
     }
   }