+++ R/Sql.R
@@ -1,6 +1,6 @@
 # @file Sql.R
 #
-# Copyright 2019 Observational Health Data Sciences and Informatics
+# Copyright 2020 Observational Health Data Sciences and Informatics
 #
 # This file is part of DatabaseConnector
 #
@@ -559,7 +559,7 @@ querySql.ffdf <- function(connection, sql, errorReportFile = file.path(getwd(), 
 #'                            executed there. This will be faster if you have many small SQL statements, but
 #'                            there will be no progress bar, and no per-statement error messages. If the 
 #'                            database platform does not support batched updates the query is executed as 
-#'                            ordinally.
+#'                            ordinarily.
 #' @param oracleTempSchema    A schema that can be used to create temp tables in when using Oracle or Impala.
 #' @param ...                 Parameters that will be used to render the SQL.
 #'