+++ NEWS.md
@@ -1,3 +1,29 @@
+DatabaseConnector 2.4.3
+=======================
+
+Changes:
+
+1. Adding bulk upload for Hive.
+
+Bugfixes:
+
+2. Reducing batch size when performing CTAS hack on Hive to avoid error.
+
+
+DatabaseConnector 2.4.2
+=======================
+
+Changes:
+
+1. BigQuery insertTable now also uses CTAS hack for improved speed and avoiding insert size limitations.
+
+Bugfixes:
+
+1. Preventing scientific notation when bulk uploading to PDW to avoid error.
+
+2. Fixing null error when calling getSchemaNames for BigQuery.
+
+
 DatabaseConnector 2.4.1
 =======================